refactor(tee,fdc2): apply pre-deploy review changes
Changes addressing findings from the TEE/FDC2 pre-deployment review,
spanning the TEE diamond facets/libraries, FDC2 modules, their public
interfaces, and the corresponding Forge tests and FCC specs.

Notable items:
- Remove the unused public `verifyAvailabilityCheckProof` query from
  VerificationFacet/IVerification (dead on-chain and off-chain); drop its
  selector from the deployer and its test block.
- Emit events from WalletResumeFacet (`PausingAddressesSet`,
  `WalletResumed`), declared on IWalletResume; assert them in tests.
- Raise the pragma floor to `^0.8.35` on every contract that
  transitively imports an `erc7201(...)` namespaced-storage library, so
  each file's stated minimum matches what it actually requires.

+++ test-forge/unit/tee/implementation/MachinePathManagerFacet.t.sol
@@ -60,7 +60,7 @@ contract TestMachinePathHelperFacet is ITestMachinePathHelper {
             owner: address(uint160(uint256(uint160(_teeId)) ^ 1)),
             teeProxyId: _teeId,
             status: _status,
-            lastStatusChangeTs: block.timestamp,
+            lastStatusChangeTs: uint64(block.timestamp),
             codeHash: _codeHash,
             platform: _platform,
             governanceHash: _governanceHash,
